+++ src/pages/index.page.tsx
@@ -1,26 +1,10 @@
-import type { NextPage } from 'next'
 import Head from 'next/head'
 
-import { styled } from '@/styles/StyleConstants'
+import BasicLayout from '../components/templates/BasicLayout'
+import Home from '../components/templates/Home'
+import type { NextPageWithLayout } from './_app.page'
 
-const Home = () => (
-  <>
-    <TopPageTitleWrapper>Welcome to roottool&apos;s portfolio!</TopPageTitleWrapper>
-    <p>&quot;Why do it yourself when robots do it better?&quot;</p>
-    <p>- echo -</p>
-  </>
-)
-
-const TopPageTitleWrapper = styled('h1', {
-  fontSize: '3.5rem',
-  marginTop: '40vh',
-  '@bp2': {
-    fontSize: '2rem',
-    marginTop: '20vh',
-  },
-})
-
-const Container: NextPage = () => (
+const Container: NextPageWithLayout = () => (
   <>
     <Head>
       <title>roottool&apos;s Portfolio Site</title>
@@ -28,5 +12,6 @@ const Container: NextPage = () => (
     <Home />
   </>
 )
+Container.getLayout = BasicLayout
 
 export default Container
