+++ src/pages/works/index.page.tsx
@@ -1,16 +1,17 @@
 import { Button, Card, CardActions, CardContent, Grid, Typography } from '@material-ui/core'
 import { createStyles, withStyles, type WithStyles } from '@material-ui/core/styles'
-import type { NextPage } from 'next'
 import Head from 'next/head'
 
 import PageTitleWrapper from '@/components/atoms/PageTitleWrapper'
 
+import BasicLayout from '../../components/templates/BasicLayout'
+import type { NextPageWithLayout } from '../_app.page'
 import contents from './WorksContents.json'
 
 const Works = ({
   classes: { card, cardGrid, container, sourceLink },
 }: WithStyles<typeof styleSettings>) => (
-  <>
+  <main>
     <PageTitleWrapper>Works</PageTitleWrapper>
     <Grid container className={container}>
       {contents.map((item, key) => (
@@ -35,7 +36,7 @@ const Works = ({
         </Grid>
       ))}
     </Grid>
-  </>
+  </main>
 )
 
 const styleSettings = () =>
@@ -57,7 +58,7 @@ const styleSettings = () =>
   })
 const StyledWorks = withStyles(styleSettings)(Works)
 
-const Container: NextPage = () => (
+const Container: NextPageWithLayout = () => (
   <>
     <Head>
       <title>Works - roottool&apos;s Portfolio Site</title>
@@ -65,5 +66,6 @@ const Container: NextPage = () => (
     <StyledWorks />
   </>
 )
+Container.getLayout = BasicLayout
 
 export default Container
